fix settings route registration
