+++ public/styles/main.css
@@ -1399,3 +1399,27 @@ header {
 #scroll-to-input.visible {
     display: flex;
 }
+
+/* Settings page tab buttons */
+.settings-tabs button {
+    background: #f5f5f5;
+    border: 1px solid #e0e0e0;
+    border-radius: 4px;
+    padding: 4px 10px;
+    cursor: pointer;
+    font-size: 12px;
+    color: #666;
+    margin-left: 8px;
+    transition: all 0.2s ease;
+}
+.settings-tabs button:hover {
+    background: #e0e0e0;
+}
+.settings-tabs button.active {
+    background: #007bff;
+    color: #fff;
+    border-color: #0056b3;
+}
+.settings-tab { display: none; }
+.settings-tab.active { display: block; }
+
